+++ flowman-tools/src/main/scala/com/dimajix/flowman/tools/exec/project/PhaseCommand.scala
@@ -61,6 +61,8 @@ sealed class PhaseCommand(phase:Phase) extends Command {
 
     override def execute(session: Session, project: Project, context:Context) : Status = {
         val args = splitSettings(this.args).toMap
+        val targets = this.targets.flatMap(_.split(","))
+        val dirtyTargets = this.dirtyTargets.flatMap(_.split(","))
         val job = "main"
         Try {
             context.getJob(JobIdentifier(job))
@@ -77,7 +79,7 @@ sealed class PhaseCommand(phase:Phase) extends Command {
                         Lifecycle.ofPhase(phase)
 
                 val coordinator = new JobCoordinator(session, force, keepGoing, dryRun, parallelism)
-                coordinator.execute(job, lifecycle, job.parseArguments(args), targets.map(_.r), dirtyTargets = dirtyTargets.map(_.r))
+                coordinator.execute(job, lifecycle, job.parseArguments(args), targets.map(_.r), dirtyTargets=dirtyTargets.map(_.r))
         }
     }
 }